// apps/portal/src/api/generated/model/createVoteQuestionDtoType.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */

export type CreateVoteQuestionDtoType =
    (typeof CreateVoteQuestionDtoType)[keyof typeof CreateVoteQuestionDtoType];

export const CreateVoteQuestionDtoType = {
    YES_NO: "YES_NO",
    SINGLE_CHOICE: "SINGLE_CHOICE",
} as const;